Minor adjustments to TopAppBarPR
getting it working slightly with the feed button for now.
