+++ src/client/app/src/main/java/com/example/ravengamingnews/MainActivity.kt
@@ -4,22 +4,14 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.DrawerValue
 import androidx.compose.material3.ModalNavigationDrawer
-import androidx.compose.material3.Scaffold
 import androidx.compose.material3.rememberDrawerState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.compose.rememberNavController
-import androidx.compose.ui.Modifier
-import androidx.navigation.compose.currentBackStackEntryAsState
-import com.example.ravengamingnews.ui.components.TopAppBarPR
 import com.example.ravengamingnews.ui.SettingsDrawer
 import com.example.ravengamingnews.ui.theme.RavenGamingNewsTheme
-import kotlinx.coroutines.launch
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -29,18 +21,6 @@ class MainActivity : ComponentActivity() {
             RavenGamingNewsTheme {
                 val navController = rememberNavController()
                 val drawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
-                val scope = rememberCoroutineScope()
-
-                val navBackStackEntry by navController.currentBackStackEntryAsState()
-                val currentDestination = navBackStackEntry?.destination?.route
-
-                val selectedTab = when (currentDestination) {
-                    "feed" -> "FEED"
-                    "all" -> "Login"
-                    "browse" -> "Login"
-                    else -> null
-                }
-
                 ModalNavigationDrawer(
                     drawerState = drawerState,
                     drawerContent = { SettingsDrawer(
@@ -48,22 +28,7 @@ class MainActivity : ComponentActivity() {
                         drawerState = drawerState,
                     ) },
                 ) {
-                    Scaffold(
-                        topBar = {
-                            TopAppBarPR(
-                                selectedTab = selectedTab,
-                                onTabSelected = { tab ->
-                                    navController.navigate(tab)
-                                },
-                                onMenuClick = {
-                                    scope.launch { drawerState.open() }
-                                }
-                            )
-                        }
-                    ) { innerPadding ->
-                        TempAppScreen(navController,
-                        modifier = Modifier.padding(innerPadding))
-                    }
+                    TempAppScreen(navController)
                 }
             }
         }
