+++ src/client/app/src/main/java/com/example/ravengamingnews/ui/components/CommonUi.kt
@@ -1,5 +1,6 @@
 package com.example.ravengamingnews.ui.components
 
+import android.view.Surface
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -24,10 +25,12 @@ import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults.mediumTopAppBarColors
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -39,6 +42,7 @@ import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.Alignment
+import androidx.navigation.NavHostController
 import com.example.ravengamingnews.R
 import com.example.ravengamingnews.ui.theme.CommonUiSize
 import com.example.ravengamingnews.ui.theme.RavenGamingNewsTheme
@@ -273,56 +277,4 @@ fun OutlinedTextFieldPreview() {
             }
         }
     }
-}
-
-@Composable
-fun TopAppBarPR(
-    selectedTab: String? = null,
-    onTabSelected: (String) -> Unit,
-    onMenuClick: () -> Unit
-) {
-    TopAppBar(
-        modifier = Modifier.fillMaxWidth().background(MaterialTheme.colorScheme.primaryContainer),
-        title = {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .background(MaterialTheme.colorScheme.primaryContainer)
-            ) {
-                // Waffle Menu
-                IconButton(onClick = onMenuClick) {
-                    Icon(
-                        imageVector = Icons.Default.Menu,
-                        contentDescription = "Menu",
-                        tint = MaterialTheme.colorScheme.onSurface
-                    )
-                }
-
-                // Navigation buttons (Feed, All, Browse)
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier = Modifier.weight(1f)
-                ) {
-                    val tabs = listOf("feed", "all", "browse")
-
-                    tabs.forEach { route ->
-                        TopAppBarButtonPR(
-                            text = route,
-                            selected = selectedTab == route,
-                            onClick = { onTabSelected(route) },
-                            modifier = Modifier.padding(horizontal = 8.dp)
-                        )
-                    }
-                }
-
-                // Logo next to menu
-                LogoImagePR(
-                    modifier = Modifier
-                        .size(80.dp)
-                        .padding(start = 8.dp, end = 16.dp)
-                )
-            }
-        }
-    )
 }
